remove useless function
